Refactored server.js file. Moved socket and production (helmet) functionality into their own modules.
