+++ modules/production.js
@@ -1,0 +1,5 @@
+const helmet = require('helmet');
+
+module.exports = function(app) {
+    app.use(helmet());
+}
